backend/requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
apify_client==2.5.0
apify_shared==2.2.0
attrs==26.1.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
cloudpathlib==0.23.0
colorama==0.4.6
confection==1.3.3
cymem==2.0.13
datasets==4.8.4
dill==0.4.1
emoji==2.15.0
fastapi==0.136.0
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.12
impit==0.12.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==11.0.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
murmurhash==1.0.15
networkx==3.6.1
numpy==2.4.4
packaging==26.1
pandas==3.0.2
preshed==3.0.13
propcache==0.4.1
psutil==7.2.2
pyarrow==24.0.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pysentimiento==0.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.6.0
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
starlette==1.0.0
sympy==1.14.0
thinc==8.3.13
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.45.0
wasabi==1.1.3
weasel==1.0.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
